split swagger generation: service and routes
