+++ src/main/resources/public/app.html
@@ -87,7 +87,7 @@
         <div class="well">
             <h1>Wasserstand: 45%</h1>
             <h1>Temperatur: 15°C</h1>
-            <h1>${message}</h1>
+
         </div>
     </section>
